Name Change for Distance Conditions
To keep a consistent naming scheme for all modifiers. This should be appropriate.
