+++ src/com/nisovin/magicspells/castmodifiers/conditions/AliveLessThanCondition.java
@@ -5,17 +5,17 @@ import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Player;
 
 import com.nisovin.magicspells.DebugHandler;
+import com.nisovin.magicspells.MagicSpells;
 import com.nisovin.magicspells.castmodifiers.Condition;
 
-public class DistanceMoreThan extends Condition {
-
-	private double distanceSq;
+	
+public class AliveLessThanCondition extends Condition {
+	int time;
 	
 	@Override
 	public boolean setVar(String var) {
 		try {
-			distanceSq = Double.parseDouble(var);
-			distanceSq = distanceSq * distanceSq;
+			time = Integer.parseInt(var);
 			return true;
 		} catch (NumberFormatException e) {
 			DebugHandler.debugNumberFormat(e);
@@ -25,18 +25,17 @@ public class DistanceMoreThan extends Condition {
 
 	@Override
 	public boolean check(Player player) {
-		return false;
+		return MagicSpells.getLifeLengthTracker().getCurrentLifeLength(player) < time;
 	}
 
 	@Override
 	public boolean check(Player player, LivingEntity target) {
-		return player.getLocation().distanceSquared(target.getLocation()) > distanceSq;
+		return target instanceof Player && check((Player)target);
 	}
 
-	
 	@Override
 	public boolean check(Player player, Location location) {
-		return player.getLocation().distanceSquared(location) > distanceSq;
+		return false;
 	}
 
 }
