+++ src/com/nisovin/magicspells/castmodifiers/conditions/DistanceMoreThanCondition.java
@@ -7,7 +7,7 @@ import org.bukkit.entity.Player;
 import com.nisovin.magicspells.DebugHandler;
 import com.nisovin.magicspells.castmodifiers.Condition;
 
-public class DistanceLessThan extends Condition {
+public class DistanceMoreThanCondition extends Condition {
 
 	private double distanceSq;
 	
@@ -30,12 +30,13 @@ public class DistanceLessThan extends Condition {
 
 	@Override
 	public boolean check(Player player, LivingEntity target) {
-		return player.getLocation().distanceSquared(target.getLocation()) < distanceSq;
+		return player.getLocation().distanceSquared(target.getLocation()) > distanceSq;
 	}
+
 	
 	@Override
 	public boolean check(Player player, Location location) {
-		return player.getLocation().distanceSquared(location) < distanceSq;
+		return player.getLocation().distanceSquared(location) > distanceSq;
 	}
 
 }
